+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityApplicationPanelBuilder, domainType.name() now used as application name by default.
 - EntityApplicationPanelBuilder.applicationVersion() renamed version().
+- EntitySearchField.Builder and EntitySearchFieldPanel.Builder.selectorFactory() renamed selector().
 
 ## 0.18.47
 ### is.codion.common.core